# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prudentide API Documentation",
    "description": "OpenAPI documentation for the Prudentide API",
    "version": "1.0",
    "contact": {
      "name": "Anders Evenrud",
      "email": "andersevenrud@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "jwt": []
    },
    {
      "auth0": []
    }
  ],
...
  "paths": {
...
    "/v1/user/{user}/avatar": {
      "get": {
        "operationId": "getUserAvatar",
        "summary": "Gets user avatar",
        "security": [],
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/getAvatar"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
    "/v1/project/{project}/webhook/{webhook}": {
      "put": {
        "operationId": "updateWebhook",
        "summary": "Update webhook",
        "tags": [
          "project webhook"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/project"
          },
          {
            "$ref": "#/components/parameters/webhook"
          },
          {
            "$ref": "#/components/parameters/conflictCheck"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/putWebhook"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/postOrPutWebhook"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      },
      "delete": {
        "operationId": "deleteWebhook",
        "summary": "Delete webhook",
        "tags": [
          "project webhook"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/project"
          },
          {
            "$ref": "#/components/parameters/webhook"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      }
    },
...
    "/v1/task/{task}/watch": {
      "get": {
        "operationId": "getTaskWatchers",
        "summary": "View project task watchers",
        "tags": [
          "task watcher"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/task"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/getTaskWatchers"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      },
      "post": {
        "operationId": "watchTask",
        "summary": "Watch project task",
        "tags": [
          "task watcher"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/task"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      },
      "delete": {
        "operationId": "unwatchTask",
        "summary": "Unwatch project task",
        "tags": [
          "task watcher"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/task"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      }
    },
...
    "/v1/task/{task}/comment/{comment}/stick": {
      "post": {
        "operationId": "stickTaskComment",
        "summary": "Sticky a task comment",
        "tags": [
          "task comment"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/task"
          },
          {
            "$ref": "#/components/parameters/comment"
          },
          {
            "$ref": "#/components/parameters/conflictCheck"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      },
      "delete": {
        "operationId": "unstickTaskComment",
        "summary": "Unsticky a task comment",
        "tags": [
          "task comment"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/task"
          },
          {
            "$ref": "#/components/parameters/comment"
          },
          {
            "$ref": "#/components/parameters/conflictCheck"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      }
    },
    "/v1/task/{task}/comment/{comment}": {
      "put": {
        "operationId": "updateTaskComment",
        "summary": "Update project task comment",
        "tags": [
          "task comment"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/task"
          },
          {
            "$ref": "#/components/parameters/comment"
          },
          {
            "$ref": "#/components/parameters/conflictCheck"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/putComment"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/postOrPutComment"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ValidationResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorResult"
          },
          {
            "type": "object",
            "properties": {
              "validation": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "keyword": {
                      "type": "string"
                    },
                    "dataPath": {
                      "type": "string"
                    },
                    "schemaPath": {
                      "type": "string"
                    },
                    "params": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "WebhookData": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "enum": [
              null,
              "post",
              "put",
              "get"
            ]
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "pattern": "^https?:"
          },
          "token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "headers": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "task:open",
                "task:close",
                "task:update",
                "task:comment"
              ]
            },
            "default": [
              "task:close"
            ]
          }
        }
      },
      "Webhook": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "projectId": {
                "type": "integer"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          {
            "$ref": "#/components/schemas/WebhookData"
          }
        ]
      },
...
      "TaskCommentData": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "sticky": {
            "type": "boolean"
          }
        }
      },
      "TaskComment": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "taskId": {
                "type": "integer"
              },
              "userId": {
                "type": "integer"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          {
            "$ref": "#/components/schemas/TaskCommentData"
          }
        ]
      },
      "TaskAttachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "taskId": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "filename": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "requestBodies": {
...
      "putWebhook": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WebhookData"
            }
          }
        }
      },
...
      "putComment": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TaskCommentData"
            }
          }
        }
      }
    },
    "responses": {
      "204": {
        "description": "OK (empty response)"
      },
      "400": {
        "description": "Invalid resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResult"
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResult"
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResult"
            }
          }
        }
      },
      "409": {
        "description": "The resource was updated after the provided reference timestamp",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResult"
            }
          }
        }
      },
      "422": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationResult"
            }
          }
        }
      },
...
      "getAvatar": {
        "description": "OK",
        "content": {
          "image/png": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
...
      "postOrPutWebhook": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Webhook"
              }
            }
          }
        }
      },
...
      "getTaskWatchers": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TaskAttachment"
              }
            }
          }
        }
      },
...
      "postOrPutComment": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TaskComment"
            }
          }
        }
      }
    },
    "parameters": {
...
      "user": {
        "name": "user",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "project": {
        "name": "project",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "webhook": {
        "name": "webhook",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "task": {
        "name": "task",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "comment": {
        "name": "comment",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "conflictCheck": {
        "name": "x-reference-timestamp",
        "in": "header",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "description": "HTTP JWT Bearer Token",
        "scheme": "bearer"
      },
      "auth0": {
        "type": "oauth2",
        "description": "Auth0 Security Token",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://os-js.eu.auth0.com/authorize?audience=https://api.prudentide.local",
            "refreshUrl": "https://os-js.eu.auth0.com/oauth/token",
            "scopes": {
              "openid": "User profile",
              "email": "Email address"
            }
          },
          "clientCredentials": {
            "tokenUrl": "https://os-js.eu.auth0.com/oauth/token",
            "refreshUrl": "https://os-js.eu.auth0.com/oauth/token",
            "scopes": {
              "openid": "User profile",
              "email": "Email address"
            }
          }
        }
      }
    }
  },
...
}